Add tabs in debug pane
