+++ src/components/viz-pane/index.css
@@ -9,3 +9,38 @@
 .Pane::-webkit-scrollbar {
   display: none;
 }
+
+.tabs-nav {
+  display: flex;
+  align-items: center;
+  justify-content: flex-start;
+  width: 100%;
+  height: 40px;
+  list-style: none;
+  margin: 0;
+  padding: 0;
+  border-bottom: 2px solid rgba(0, 0, 0, 0.08);
+}
+
+.tabs-nav li {
+  display: flex;
+  align-items: center;
+  height: 100%;
+  float: left;
+  padding: 0 10px;
+}
+
+.active-tab {
+  background: rgba(0, 0, 0, 0.08);
+}
+
+.debug-pane {
+  width: 100%;
+}
+
+.close {
+  position: absolute;
+  top: 10px;
+  right: 15px;
+  cursor: pointer;
+}
