+++ src/components/viz-pane/renderer.tsx
@@ -1,8 +1,10 @@
 import './index.css';
 
 import * as React from 'react';
+import { X } from 'react-feather';
 import SplitPane from 'react-split-pane';
 
+import DataViewer from '../data-viewer';
 import ErrorBoundary from '../error-boundary';
 import ErrorPane from '../error-pane';
 import Renderer from '../renderer';
@@ -10,9 +12,21 @@ import Toolbar from '../toolbar';
 
 interface Props {
   errorPane?: boolean;
+
+  showErrorPane: () => void;
 }
 
-export default class VizPane extends React.Component<Props> {
+interface State {
+  errorLogs: boolean;
+}
+
+export default class VizPane extends React.Component<Props, State> {
+  constructor(props) {
+    super(props);
+    this.state = {
+      errorLogs: true,
+    };
+  }
   public render() {
     const container = (
       <div className="chart-container">
@@ -26,7 +40,26 @@ export default class VizPane extends React.Component<Props> {
       return (
         <SplitPane split="horizontal" defaultSize={window.innerHeight * 0.6} paneStyle={{ display: 'flex' }}>
           {container}
-          <ErrorPane />
+          <div className="debug-pane">
+            <span onClick={e => this.props.showErrorPane()} className="close">
+              <X />
+            </span>
+            <ul className="tabs-nav">
+              <li
+                className={this.state.errorLogs ? 'active-tab' : ''}
+                onClick={() => this.setState({ errorLogs: true })}
+              >
+                Error Logs
+              </li>
+              <li
+                className={this.state.errorLogs ? '' : 'active-tab'}
+                onClick={() => this.setState({ errorLogs: false })}
+              >
+                Data Viewer
+              </li>
+            </ul>
+            {this.state.errorLogs ? <ErrorPane /> : <DataViewer />}
+          </div>
         </SplitPane>
       );
     } else {
